LayoutTests/inspector/unit-tests/text-utilities-expected.txt

Testing basic functionality of functions defined in TextUtilities.js.


== Running test suite: TextUtilities
-- Running test case: WebInspector.rangeForNextCSSNameOrValue
PASS: Next name/value token in "" starting at index 0 is "" [0, 0]
PASS: Next name/value token in "" starting at index 2 is "" [0, 0]
PASS: Next name/value token in "foo" starting at index 0 is "foo" [0, 3]
PASS: Next name/value token in "foo" starting at index 2 is "foo" [0, 3]
PASS: Next name/value token in "foo:bar" starting at index 0 is "foo" [0, 3]
PASS: Next name/value token in "foo:bar" starting at index 2 is "foo" [0, 3]
PASS: Next name/value token in "foo:bar" starting at index 3 is "bar" [4, 7]
PASS: Next name/value token in "foo:bar" starting at index 5 is "bar" [4, 7]
PASS: Next name/value token in "foo:  bar  ;" starting at index 0 is "foo" [0, 3]
PASS: Next name/value token in "foo:  bar  ;" starting at index 2 is "foo" [0, 3]
PASS: Next name/value token in "foo:  bar  ;" starting at index 3 is "bar" [6, 9]
PASS: Next name/value token in "foo:  bar  ;" starting at index 5 is "bar" [6, 9]
PASS: Next name/value token in "foo: url(http://baz);" starting at index 0 is "foo" [0, 3]
PASS: Next name/value token in "foo: url(http://baz);" starting at index 2 is "foo" [0, 3]
PASS: Next name/value token in "foo: url(http://baz);" starting at index 3 is "url(http://baz)" [5, 20]
PASS: Next name/value token in "foo: url(http://baz);" starting at index 5 is "url(http://baz)" [5, 20]
